add ComponentContainer for IORequestCRTP
