+++ modules/runtime/src/io/components/file_component.hpp
@@ -1,0 +1,39 @@
+#pragma once
+#include "SkrRT/io/io.h"
+#include "SkrRT/platform/dstorage.h"
+#include "SkrRT/platform/guid.hpp"
+#include "../components/component.hpp"
+
+namespace skr {
+namespace io {
+
+template <>
+struct IORequestComponentTID<struct IORequestFile> 
+{
+    static constexpr skr_guid_t Get()
+    {
+        using namespace skr::guid::literals;
+        return u8"6bf19e92-7180-42d5-9bb7-19cae4e8716d"_guid;
+    } 
+};
+struct IORequestFile : public IORequestComponent
+{
+    IORequestFile(IIORequest* const request) SKR_NOEXCEPT 
+        : IORequestComponent(request) 
+    {
+        
+    }
+    virtual skr_guid_t get_tid() const SKR_NOEXCEPT override { return IORequestComponentTID<IORequestFile>::Get(); }
+
+    void set_vfs(skr_vfs_t* _vfs) SKR_NOEXCEPT { vfs = _vfs; }
+    void set_path(const char8_t* p) SKR_NOEXCEPT { path = p; }
+    const char8_t* get_path() const SKR_NOEXCEPT { return path.u8_str(); }
+
+    skr::string path;
+    skr_vfs_t* vfs = nullptr;
+    skr_io_file_handle file = nullptr;
+    SkrDStorageFileHandle dfile = nullptr;
+};
+
+} // namespace io
+} // namespace skr
